pyproject.toml
[project]
name = "whirl"
version = "0.1.0"
description = "Simple orchestration tool for ipynb notebooks"
authors = [
    {name = "Vojtech Mikes"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (==6.0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)"
]
